Revert from (Add initial config in jest.config.js) to (Add type to PrivateRoute FC)
Reason: the current build(CRA-jest) does not need these changes. These changes were applied because of a problem running the tests with 'npm test'. However these changes are not the right solution with the current configuration provided by CRA. Due to the problem, Jest can't find any of the tests, returning exit code 0. This problem will be fixed in the next commit.