Fix 'No tests found, exiting with code 1'
This changes the default configuration of 'react-scripts test' to find the tests under 'src' folder. If this change is not applied then the default config of react-scripts will search for test files considering the name of the parent directories of the project. This change prevent tests from not being found when the test command is run if there is hidden or named parent directory, such as '/.XYZ/'